Sacar casteo a int.

+++ parser/parser_inventario.cpp
@@ -20,9 +20,8 @@ void ParserInventario::parse(const std::string& input,
 	std::regex pattern("(\\w+) (\\d+) (\\d+)");
 	std::smatch match;
 	std::regex_search(input, match, pattern);
-
-	Material p1 (nombre(match), (int)cantidad(match, 1));
+	Material p1 (nombre(match), cantidad(match, 1));
 	inventario_p1.alta_al_final(p1);
-	Material p2 (nombre(match), (int)cantidad(match, 2));
+	Material p2 (nombre(match), cantidad(match, 2));
 	inventario_p2.alta_al_final(p2);
 }
